IMR-69 refactor: change youtube_crawler to keyCrawler
